Use "Long" data type for chatId fields

+++ library/src/main/java/com/pengrad/telegrambot/request/CloseForumTopic.java
@@ -3,7 +3,7 @@ package com.pengrad.telegrambot.request;
 import com.pengrad.telegrambot.response.BaseResponse;
 
 public class CloseForumTopic extends BaseRequest<CloseForumTopic, BaseResponse> {
-    public CloseForumTopic(Integer chatId, Integer messageThreadId) {
+    public CloseForumTopic(Long chatId, Integer messageThreadId) {
         super(BaseResponse.class);
         add("chat_id", chatId);
         add("message_thread_id", messageThreadId);
