+++ library/src/main/java/com/pengrad/telegrambot/request/DeleteForumTopic.java
@@ -3,7 +3,7 @@ package com.pengrad.telegrambot.request;
 import com.pengrad.telegrambot.response.BaseResponse;
 
 public class DeleteForumTopic extends BaseRequest<DeleteForumTopic, BaseResponse> {
-    public DeleteForumTopic(Integer chatId, Integer messageThreadId) {
+    public DeleteForumTopic(Long chatId, Integer messageThreadId) {
         super(BaseResponse.class);
         add("chat_id", chatId);
         add("message_thread_id", messageThreadId);
